Move internal items to internal folder and adjust namespaces accordingly
