+++ Neo4j.Driver/Neo4j.Driver.Tests/Mapping/BuiltMapperTests.cs
@@ -14,6 +14,7 @@
 // limitations under the License.
 
 using FluentAssertions;
+using Neo4j.Driver.Internal.Mapping;
 using Neo4j.Driver.Mapping;
 using Neo4j.Driver.Tests.TestUtil;
 using Xunit;
