+++ Neo4j.Driver/Neo4j.Driver.Tests/Mapping/DefaultMapperTests.cs
@@ -16,6 +16,7 @@
 using System;
 using System.Collections.Generic;
 using FluentAssertions;
+using Neo4j.Driver.Internal.Mapping;
 using Neo4j.Driver.Internal.Types;
 using Neo4j.Driver.Mapping;
 using Neo4j.Driver.Tests.TestUtil;
