+++ Neo4j.Driver/Neo4j.Driver.Tests/Mapping/MappableValueProviderTests.cs
@@ -17,6 +17,7 @@ using System.Collections.Generic;
 using FluentAssertions;
 using Moq;
 using Moq.AutoMock;
+using Neo4j.Driver.Internal.Mapping;
 using Neo4j.Driver.Internal.Types;
 using Neo4j.Driver.Mapping;
 using Xunit;
