+++ Neo4j.Driver/Neo4j.Driver.Tests/Mapping/MappedListCreatorTests.cs
@@ -18,6 +18,7 @@ using System.Linq;
 using FluentAssertions;
 using Moq;
 using Moq.AutoMock;
+using Neo4j.Driver.Internal.Mapping;
 using Neo4j.Driver.Mapping;
 using Neo4j.Driver.Tests.TestUtil;
 using Xunit;
