+++ Neo4j.Driver/Neo4j.Driver.Tests/Mapping/MappingSourceDelegateBuilderTests.cs
@@ -15,6 +15,7 @@
 
 using System.Collections.Generic;
 using FluentAssertions;
+using Neo4j.Driver.Internal.Mapping;
 using Neo4j.Driver.Internal.Types;
 using Neo4j.Driver.Mapping;
 using Neo4j.Driver.Tests.TestUtil;
