+++ Neo4j.Driver/Neo4j.Driver/Internal/Mapping/BuiltMapper.cs
@@ -17,8 +17,9 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
+using Neo4j.Driver.Mapping;
 
-namespace Neo4j.Driver.Mapping;
+namespace Neo4j.Driver.Internal.Mapping;
 
 internal class BuiltMapper<T> : IRecordMapper<T>
 {
