+++ Neo4j.Driver/Neo4j.Driver/Internal/Mapping/DefaultMapper.cs
@@ -16,8 +16,9 @@
 using System;
 using System.Collections.Generic;
 using System.Reflection;
+using Neo4j.Driver.Mapping;
 
-namespace Neo4j.Driver.Mapping;
+namespace Neo4j.Driver.Internal.Mapping;
 
 internal static class DefaultMapper
 {
