+++ Neo4j.Driver/Neo4j.Driver/Internal/Mapping/MappedListCreator.cs
@@ -16,8 +16,9 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
+using Neo4j.Driver.Mapping;
 
-namespace Neo4j.Driver.Mapping;
+namespace Neo4j.Driver.Internal.Mapping;
 
 internal interface IMappedListCreator
 {
