+++ Neo4j.Driver/Neo4j.Driver/Internal/Mapping/MappingSourceDelegateBuilder.cs
@@ -13,7 +13,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Neo4j.Driver.Mapping;
+using Neo4j.Driver.Mapping;
+
+namespace Neo4j.Driver.Internal.Mapping;
 
 internal delegate bool MappingValueDelegate(IRecord record, out object value);
