+++ Neo4j.Driver/Neo4j.Driver/Neo4j.Driver.csproj
@@ -51,4 +51,7 @@
     <ItemGroup Condition="'$(TargetFramework)' == 'netstandard2.0'">
         <PackageReference Include="Microsoft.Bcl.AsyncInterfaces" Version="6.0.0" />
     </ItemGroup>
+    <ItemGroup>
+      <Folder Include="Internal\Mapping\" />
+    </ItemGroup>
 </Project>
